Additional documentation around extractors
This is pretty tough to do with any kind of precision right now.  It's clear that the underlying API is in flux as new objects to handle metadata are being added.  This is meant as a placeholder right now to let people know about the fact that extractors exist so they don't spend a bunch of time trying to figure out why it doesn't work out of the box

Signed-off-by: Bob Briski <rbriski@gmail.com>